Add files via upload
This  Shiny app may help you in understanding the data from SCC2022 competion. 
Description of the challenge 
Case Study #1: Developing a physician performance model in critical care – Assessing quality and value
Patients admitted to the ICU are the sickest and most complex in the health care system requiring expensive lifesaving technologies and interventions. Further, care within an intensive care unit is provided by a collaborative team of health professionals. Many efforts have been made to develop ICU metrics to reflect the performance of a healthcare system. Traditionally, physicians participate in performance evaluations that are called 360 evaluations. The 360 model of feedback utilizes information from self-assessment, colleagues, non-physicians and patients. These assessments are useful, but fail to consider additional sources of data (e.g. patient specific outcomes including ICU or hospital length of stay, complications, and mortality) that may provide a more encompassing picture of physician performance. We are interested in developing a model for assessing the performance of an individual physician.

Link: https://ssc.ca/en/case-study/developing-a-physician-performance-model-critical-care-assessing-quality-and-value